+++ Source/Game/Wavefront.h
@@ -13,7 +13,7 @@ namespace tankwars {
         std::vector<glm::vec3> positions;
         std::vector<glm::vec2> texCoords;
         std::vector<glm::vec3> normals;
-        std::vector<uint16_t> triangles;
+        std::vector<uint32_t> triangles;
     };
     
     WavefrontModel readWavefrontFromFile(const std::string& path, bool revertWindings = false);
